requirements.txt
scikit-fem>=8.0.0
numpy>=1.26.0
scipy>=1.11.0
matplotlib>=3.8.0
PyQt5>=5.15.10
nidaqmx>=1.4.1
pyserial>=3.5
